created a class with private variables
i decided to try and create a class with private variables so i could
save some data i needed for the availableMoves method. Its not recursive
but it will work for any state of the board. Give it a look and let me
know what you think. Also added a test2 method in main to test the new
class.