added the evaluation function
I haven't gotten the chance to test it yet cause I need to add more
methods for the ai side of the game. Im still working on that on my
side. Still having problems with available moves not showing all the
moves. Theres also a bug that when you get to a point in the game it
gives you incorrect moves and if you use one of them the program breaks.
Will look more into it tomorrow after work.